thung rac 04/24

+++ src/main/java/com/poly/admin/controller/CategoryAController.java
@@ -3,7 +3,6 @@ package com.poly.admin.controller;
 import java.util.Optional;
 
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -13,12 +12,11 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 import com.poly.dao.CategoryDAO;
-
 import com.poly.entity.CategoryEntity;
-import com.poly.entity.OrderEntity;
 import com.poly.model.AccountModel;
 import com.poly.utils.SessionUtil;
 
+//----------------------------------------------------hypercare
 @Controller
 @RequestMapping("admin/category") //@RequestMapping phần chung g
 public class CategoryAController {
@@ -27,13 +25,17 @@ public class CategoryAController {
 	@Autowired
 	HttpServletRequest request;
 	
-	@RequestMapping("index")//@RequestMapping phần riêng
-	public String index(Model model) {
+	public void  USER_LOGIN(Model model) {
 		AccountModel accountModel = (AccountModel) SessionUtil.getInstance().getValue(request, "USER_LOGIN");
 		if (accountModel != null) {
 			model.addAttribute("userLogin", accountModel.getUsername());
 			model.addAttribute("role", accountModel.getRole());
 		}
+	}
+	
+	@RequestMapping("index" )//@RequestMapping phần riêng
+	public String index(Model model) {
+		USER_LOGIN(model);
 		model.addAttribute("form", new CategoryEntity());
 		model.addAttribute("list", cdao.fillAllCategoryStatus1());
 		return "admin/category/index";
@@ -42,20 +44,23 @@ public class CategoryAController {
 	@RequestMapping("edit/{id}")
 	public String edit(Model model, @PathVariable("id") long id) {
 		model.addAttribute("form", cdao.getOne(id));
-		model.addAttribute("list", cdao.findAll());
+		model.addAttribute("list", cdao.fillAllCategoryStatus1());
 		return "admin/category/index";
 	}
 	
 	@RequestMapping("create")
-	public String create(Model model, @ModelAttribute("form") CategoryEntity entity) { 
-		try {
+	public String create(Model model,@ModelAttribute("form") CategoryEntity entity) { 
+		try {			
+			entity.setStatus(1);
 			cdao.save(entity);
 			model.addAttribute("form", new CategoryEntity());
 			model.addAttribute("message", "Tạo mới loại hàng thành công!");
 		} catch (Exception e) {
 			model.addAttribute("message", "Tạo mới loại hàng thất bại!");
+			e.printStackTrace();
+			System.out.println(">>>>>>>>>>>>>>>>>>>>>>>>" +e);
 		}
-		model.addAttribute("list", cdao.findAll());
+		model.addAttribute("list", cdao.fillAllCategoryStatus1());
 		return "admin/category/index";
 	}
 	
@@ -65,10 +70,12 @@ public class CategoryAController {
 			model.addAttribute("message", "Loại hàng không tồn tại!");
 		}
 		else {
+			entity.setStatus(1);
 			cdao.save(entity);
+			model.addAttribute("form", new CategoryEntity());
 			model.addAttribute("message", "Cập nhật loại hàng thành công!");
 		}
-		model.addAttribute("list", cdao.findAll());
+		model.addAttribute("list", cdao.fillAllCategoryStatus1());
 		return "admin/category/index";
 	}
 	
@@ -79,11 +86,84 @@ public class CategoryAController {
 			model.addAttribute("message", "Loại hàng không tồn tại!");
 		}
 		else {
-			cdao.delete(option.get());
+			entity.setStatus(2);
+			cdao.save(entity);
 			model.addAttribute("form", new CategoryEntity());
 			model.addAttribute("message", "Xóa loại hàng thành công!");
 		}
-		model.addAttribute("list", cdao.findAll());
+		model.addAttribute("list", cdao.fillAllCategoryStatus1());
 		return "admin/category/index";
 	}
+	
+	/**/
+	@RequestMapping("edit/index")//@RequestMapping phần riêng
+	public String editindex(Model model) {
+		USER_LOGIN(model);
+		model.addAttribute("form", new CategoryEntity());
+		model.addAttribute("list", cdao.fillAllCategoryStatus1());
+		return "admin/category/index";
+	}
+	@RequestMapping("edit/_thungrac")//@RequestMapping phần riêng
+	public String editindex_thungrac(Model model) {
+		USER_LOGIN(model);
+		model.addAttribute("formt", new CategoryEntity());
+		model.addAttribute("listfillAllCategoryStatus2", cdao.fillAllCategoryStatus2());
+		return "admin/category/_thungrac";
+	}
+	
+	@RequestMapping("_thungrac")//@RequestMapping phần riêng
+	public String index_thungrac(Model model) {
+		USER_LOGIN(model);
+		model.addAttribute("formt", new CategoryEntity());
+		model.addAttribute("listfillAllCategoryStatus2", cdao.fillAllCategoryStatus2());
+		return "admin/category/_thungrac";
+	}
+	
+//	@RequestMapping("update2")
+//	public String update2(Model model, @ModelAttribute("form") CategoryEntity entity) {
+//		if(!cdao.existsById(entity.getId())) { 
+//			model.addAttribute("message", "Loại hàng không tồn tại!");
+//		}
+//		else {
+//			entity.setStatus(1);
+//			cdao.save(entity);
+//			model.addAttribute("form", new CategoryEntity());
+//			model.addAttribute("message", "Lấy lại loại hàng thành công!");
+//		}
+//		model.addAttribute("list", cdao.fillAllCategoryStatus1());
+//		return "admin/category/index";
+//	}
+	
+	@RequestMapping("laylai/{id}/{name}")
+	public String laylai_thungrac(Model model, @PathVariable("id") long id ,CategoryEntity entity) {
+		if(!cdao.existsById(entity.getId())) { 
+			model.addAttribute("message", "Loại hàng không tồn tại!");
+		}
+		else {
+			entity.setStatus(1);
+			cdao.save(entity);
+			model.addAttribute("message", "Lấy lại loại hàng thành công!");
+		}
+		model.addAttribute("listfillAllCategoryStatus2", cdao.fillAllCategoryStatus2());
+		return "admin/category/_thungrac";
+	}
+	
+	
+	@RequestMapping("delete/{id}")// xóa bị lỗi
+	public String delete1(Model model,@PathVariable("id") long id, CategoryEntity entity) {
+		Optional<CategoryEntity> option = cdao.findById(entity.getId());
+		if(!option.isPresent()) {
+			model.addAttribute("message", "Loại hàng không tồn tại!");
+		}
+		else {
+			cdao.delete(option.get());
+		
+			model.addAttribute("message", "Xóa loại hàng thành công!");
+		}
+		model.addAttribute("listfillAllCategoryStatus2", cdao.fillAllCategoryStatus2());
+		return "admin/category/_thungrac";
+	}
+
+	/**/
+	
 }
